+++ components/gnb/gnb.tsx
@@ -1,96 +1,17 @@
-'use client';
-
 import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 
-import NotificationBtn from './notificationBtn';
+import AuthButton from './authButton';
 import SearchBar from './searchBar';
 
-// 더미데이터
-const alertList = {
-  limit: 5, // 알림 개수
-  items: [
-    {
-      item: {
-        id: '1',
-        createdAt: '2024-04-18T08:00:00Z',
-        result: 'accepted', // 또는 'rejected'
-        shop: {
-          item: {
-            id: '1',
-            name: 'Shop A',
-            imageUrl: 'https://example.com/shop_a_image.jpg',
-          },
-        },
-        notice: {
-          item: {
-            startsAt: '2024-04-18T09:00:00Z',
-            workhour: 8,
-          },
-        },
-      },
-    },
-    {
-      item: {
-        id: '2',
-        createdAt: '2024-04-17T12:00:00Z',
-        result: 'rejected', // 또는 'accepted'
-        shop: {
-          item: {
-            id: '2',
-            name: 'Shop B',
-            imageUrl: 'https://example.com/shop_b_image.jpg',
-          },
-        },
-        notice: {
-          item: {
-            startsAt: '2024-04-17T13:00:00Z',
-            workhour: 6,
-          },
-        },
-      },
-    },
-  ],
-};
-interface UserTypeProps {
-  userType: string;
-}
-
-function UserTypeLink({ userType }: UserTypeProps) {
-  return userType === 'employee' ? <Link href="/my-profile">내 프로필</Link> : <Link href="/my-store">내 가게</Link>;
-}
-
 function Gnb() {
-  const router = useRouter();
-
-  const handleLogout = () => {
-    document.cookie = 'accessToken=; Path=/; Expires=Thu, 01 Jan 1970 00:00:01 GMT;';
-    router.push('/');
-  };
-
-  const auth: boolean = true;
-  const userType: string = 'employee';
   return (
     <header className="sticky top-0 z-10 px-5 py-3 m-auto bg-white md:px-8 w-full md:max-w-[964px]">
       <nav className="flex flex-wrap items-center gap-4 md:relative md:gap-10">
         <Link href="/">
           <Image src="/icons/logo.svg" width={112} height={40} alt="로고" />
         </Link>
-        <div className="flex items-center gap-5 ml-auto font-bold text-black md:order-1">
-          {auth ? (
-            <>
-              <UserTypeLink userType={userType} />
-              <button onClick={handleLogout}>로그아웃</button>
-              <NotificationBtn alertList={alertList} />
-            </>
-          ) : (
-            <>
-              <Link href="/signin">로그인</Link>
-              <Link href="/signup">회원가입</Link>
-            </>
-          )}
-        </div>
+        <AuthButton />
         <SearchBar />
       </nav>
     </header>
